Agregado mas controladores.

+++ app/Http/Controllers/AsientoController.php
@@ -36,7 +36,10 @@ class AsientoController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $asiento = new Asiento;
+        $asiento->fill($request->all());
+        $asiento->save();
+        return $asiento;
     }
 
     /**
@@ -48,7 +51,12 @@ class AsientoController extends Controller
     public function show($id)
     {
         $asiento = Asiento::find($id);
-        return $asiento;
+        if($asiento != NULL){
+            return $asiento;
+        }
+        else{
+            return "Asiento no existe.";
+        }
     }
 
     /**
@@ -71,7 +79,15 @@ class AsientoController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $asiento = Asiento::find($id);
+        if($asiento != NULL){
+          $asiento->fill($request->all());
+          $asiento->save();
+          return $asiento;
+        }
+        else{
+          return "No puede modificar un asiento no existente.";
+        }
     }
 
     /**
@@ -83,7 +99,12 @@ class AsientoController extends Controller
     public function destroy($id)
     {
       $asiento = Asiento::find($id);
-      $asiento->delete();
-      return Asiento::All();
+      if($asiento != NULL){
+          $asiento->delete();
+          return "Asiento eliminado.";
+      }
+      else{
+          return "Asiento no existe.";
+      }
     }
 }
